requirements.txt
Flask
Flask-CORS
Flask-SQLalchemy

pyjwt